requirements.txt
python-docx
IPython
boto3
pyautogen
GitPython
